square OAuth login and site and snippet list
Using a forked version of omniauth-square to allow a common version of
faraday to coexist with square.rb gem. This also allows for rails 3 to
be used.

The idea of Devise User being associated with Identity model was copied
from
  https://willschenk.com/articles/2015/setting-up-devise-with-twitter-and-facebook-and-other-omniauth-schemes-without-email-addresses/

Also note this need production credentials as snippets is only available
in production.