move single site redirect to index
only when visiting the site through the root path ro logging in should
we force the user to landing on their "only available site". At no time
should we prevent them from vieiwing all their sites. The redirect hence
needs to happen on the index action and NOT on the show action